create-and-submit modal only when neccessary
